+++ nengo/utils/tests/test_progress.py
@@ -6,9 +6,9 @@ from nengo.utils.progress import (
 
 
 class ProgressBarMock(ProgressBar):
-    def __init__(self):
+    def __init__(self, task="Testing"):
         self.n_update_calls = 0
-        super(ProgressBarMock, self).__init__()
+        super(ProgressBarMock, self).__init__(task)
 
     def update(self, progress):
         self.n_update_calls += 1
@@ -103,7 +103,7 @@ class TestUpdateN(object):
         progress_bar = ProgressBarMock()
         updater = UpdateN(progress_bar, max_updates=3)
 
-        with ProgressTracker(100, updater) as p:
+        with ProgressTracker(100, updater, "Testing") as p:
             for _ in range(100):
                 p.step()
 
@@ -116,7 +116,7 @@ class TestUpdateEveryN(object):
         progress_bar = ProgressBarMock()
         updater = UpdateEveryN(progress_bar, every_n=5)
 
-        with ProgressTracker(100, updater) as p:
+        with ProgressTracker(100, updater, "Testing") as p:
             progress_bar.n_update_calls = 0
             for _ in range(5):
                 p.step()
@@ -135,7 +135,7 @@ class TestUpdateEveryT(object):
         t = 1.
         monkeypatch.setattr(time, 'time', lambda: t)
 
-        with ProgressTracker(100, updater) as p:
+        with ProgressTracker(100, updater, "Testing") as p:
             p.step()  # Update is allowed to happen on first step.
 
             progress_bar.n_update_calls = 0